Handle IO timeouts for TLS setup properly, make badcmds count a Limits
Previously we completely fumbled IO timeouts for TLS setup and for TLS
connections, because we left previously-set read and write timeouts in
place on the underlying connection. Now we set an explicit timeout for
TLS setup then clear the timeout on the underlying connection once TLS
is up.

(Our regular code will then apply timeouts to the TLS version of the
connection.)

While I was there, I made how many bad commands we'd accept before an
abort into a Limits field.