sinksmtp.go: add another message name hash option and rename things a chunk
Motivated by systems that appear to incorrectly resend messages after
they are 5xx rejected after post-DATA (ie, -M), I've added a hash type
that is the message data plus most but not all metadata; it excludes the
ID and the source port. This resulted in reformatting the 'id ' line to
include the source port just in case we ever want to recover it.

Since sinksmtp now has three options for the message save hash, we now
set it up via a '-save-hash TYPE' command line option instead of the
old boolean flag for what is now '-save-hash msg'.