sinksmtp: switch to better handling for 'match RCPT TO, defer later' rules
This follows the scheme outlined in the aside in the previous commit.
At post-pRto time we evaluate any rule that requires pRto (or higher)
once for each (accepted) RCPT TO and take it if any of them match.
This simplies the code and throws out random bits and pieces.

As a side change, pData and pMessage now explicitly blank out
Context.rcptto's value, since it isn't well defined in general in those
phases (before, it was 'the last RCPT TO that was looked at').
This is not perfect, since checking a pRto rule with our special
hack will set Context.rcptto as a side effect and not clear it,
but it's better than nothing.