Add support for formatted, multi-line rejections and tempfails
This adds Conn.RejectMsg() and Conn.TempfailMsg(), which are probably
misnamed by Go conventions but 'Rejectf()' and 'Tempfailf()' are too close
to Reject() and Tempfail() for me. They take a format and arguments and
you can make a multiline reply with embedded newlines. You do not supply
the code; that is determined automatically.