sinksmtp: add 'set-with' rule action that just sets 'with ...' values
A set-with rule does not terminate rule checking; it's purely there to
set up 'default' values for its 'with ...' options. These values can be
overridden later by another set-with or by the non-set-with rule that
matches. This can unfortunately result in some confusing and arcane
results, but that's life; I can't think of any better semantics right
now.

(I'm reluctant to add some sort of conditional/switch-based matching
to set-with or in general, although in the abstract the idea is vaguely
tempting. The devil would be in the details and it would push condition
rules further towards being a 'real language' in some sense.)