sinksmtp: support multiple TLS keys and apparent host names
Sinksmtp can now pick TLS keys and the local host name it announces in
greeting banners based on the IP that people have connected to. This
is not completely well documented and the name I've given it and the
command line flag are probably bad. Names are not my strong thing.

Also, needs tests, but there turn out to be problems with that due
to the interface I chose.