smtpd.go: expose full TLS connection state in Conn struct, log ServerName
The connection state has enough useful things in it that we should
expose it all rather than doing so piecemeal. This is an API change
since we're removing the TLSCipher field (given that it's just copied
from tls.ConnectionState).

We also log the server name that the remote end sent us on TLS, if
any. Unsurprisingly, it seems that a lot of MTAs don't use SNI and
so this is blank.