smtpd.go: make a connection's configuration public
Conn.Config is the live configuration of the connect, which allows you
to manipulate it (for example, to delete TLS) after a Conn is created.
This turns out to be convenient for some stuff.

There are obvious restrictions that apply, eg trying to remove TLS
after STARTTLS is going to be what they call 'ineffective'.