sinksmtp: rethink TLS protocol version handling
Now we log the protocol version that was negotiated but do not insist on
a minimum version. The handling we may want for minimum TLS versions and
so on is complicated enough that it should be pushed to the configuration
language, and in the mean time having SSLv3 beats not having SSL at all.