sinksmtp: add new rule operator 'source HPAT' to match several things at once
This is equivalent to '(host HPAT or ehlo HPAT or from @HPAT)' because
frankly I keep writing rules with such a collection of or-rules so I
decided to make it convenient.

Actually I just realized that this first cut version doesn't work if
you provide a file. I'll have to do something about that that involves
a more complicated approach.