sinksmtp: explicitly run lexer to EOF on parser errors to end goroutine
The lexer does its actual lexing in a separate goroutine, which means
that if we just abandon lexing and parsing a set of rules part way
through we'll wind up with a lingering goroutine; one case where this
can happen is if there are parse errors. Now we explicitly pull items
from the lexer channel until it closes, and we close it at the end of
the lexer goroutine loop.

This fix is taken from a similar fix made to the Go standard library's
template lexing and parsing code. My test for it isn't as good as theirs,
though. That would take some internal restructuring.