Make network write errors during EHLO replies abort immediately
We previously would blindly write all lines of an EHLO reply even if
we got an immediate network error (because the client has disconnected,
which clients do do here). Now there is a crude workaround to stop
immediately.

(This is mostly noticeable if you have SMTP logging enabled; otherwise
it hardly matters.)