Delete cmd/*'s version of sinksmtp because it's obsolete
The real version of sinksmtp is in a separate repo,
	https://github.com/siebenmann/sinksmtp

I updated the README and Makefile in the last commit but didn't
delete the cmd/* files for no good reason. Well, now I have done
that. It's gone.

(There will probably be a simple sample command later, but for
now this will clean up various things that scan repos.)