Don't log EOF error on sudden connection close
If a client closes a connection before the `QUIT` command has been issued an EOF error will be reported when logging has been enabled like so:

```
! command abort 0 bytes read err: EOF
```

Since there appear to be a lot of clients out there who behave like this, catch the io.EOF error and silently close it. The `smtpd.ABORT` event will still be sent when this happens if you want to catch these situations on the implementaton side of things.