Correctly handle data from clients that contains '%'
We were previously passing data from clients as the 'format' parameter
for some printf-like methods that took a format and arguments for it.
This is incorrect and dangerous; it could result in corrupted logs and
other problems. Fix it by always passing a constant format string,
even if the constant format string is just "%s".